sticker_gen/requirements.txt
# FlockBlocker Sticker Generator — Core Dependencies
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.24.0

# Optional: EOT adversarial patch generation (not required for core strategies)
# torch>=2.1.0
